settings: move all types to their own files
Registry was getting too big and cumbersome.
